decimal changed to double in plus4

+++ MathAgain/MathAgain/Math.cs
@@ -13,9 +13,9 @@ namespace MathAgain
             int plus2 = inputX + 2;
             return plus2;
         }
-        public static int DoMath(Decimal decimalX)
+        public static int DoMath(Double decimalX)
         {
-            decimal plus4 = decimalX + 4;
+            Double plus4 = decimalX + 4;
             return Convert.ToInt16(plus4);
         }
